pyproject.toml
[tool.poetry]
name = "nft-loans"
version = "0.1.0"
description = ""
authors = ["readmost <rauan.amangeldiyev@nu.edu.kz>"]
readme = "README.md"
packages = [{include = "nft_loans"}]

[tool.poetry.dependencies]
python = "^3.9"
celery = "5.2.7"
aiohttp = "3.8.3"
aiosignal = "1.3.1"
amqp = "5.1.1"
appnope = "0.1.3"
asgiref = "3.6.0"
asn1crypto = "1.5.1"
astroid = "2.14.2"
asttokens = "2.2.1"
async-timeout = "4.0.2"
attrs = "22.2.0"
backcall = "0.2.0"
billiard = "3.6.4.0"
cachetools = "5.2.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
cloud-sql-python-connector = "1.0.0"
cryptography = "39.0.0"
decorator = "5.1.1"
dill = "0.3.6"
django = "4.1.6"
executing = "1.2.0"
frozenlist = "1.3.3"
google-auth = "2.15.0"
idna = "3.4"
ipython = "8.8.0"
isort = "5.12.0"
jedi = "0.18.2"
kombu = "5.2.4"
lazy-object-proxy = "1.9.0"
matplotlib-inline = "0.1.6"
mccabe = "0.7.0"
multidict = "6.0.4"
parso = "0.8.3"
pexpect = "4.8.0"
pg8000 = "1.29.4"
pickleshare = "0.7.5"
platformdirs = "3.0.0"
prompt-toolkit = "3.0.36"
psycopg2-binary = "2.9.5"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pygments = "2.14.0"
pylint-django = "2.5.3"
pylint-plugin-utils = "0.7"
python-dateutil = "2.8.2"
python-dotenv = "0.21.1"
pytz = "2022.7.1"
requests = "2.28.1"
rsa = "4.9"
scramp = "1.4.4"
six = "1.16.0"
sqlparse = "0.4.3"
stack-data = "0.6.2"
tomli = "2.0.1"
tomlkit = "0.11.6"
traitlets = "5.8.0"
typing-extensions = "4.4.0"
urllib3 = "1.26.13"
vine = "5.0.0"
wcwidth = "0.2.6"
wrapt = "1.14.1"
yarl = "1.8.2"
pre-commit = "^3.0.4"
black = "^23.1.0"
memcache = "^0.12.0"
django-celery-results = "^2.4.0"
redis = "^4.5.1"
django-celery-beat = "^2.4.0"
flower = "^1.2.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
ortools = "^9.5.2237"
matplotlib = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
use-single-quotes = true
